Re-arranged directory structure
